analyzer: Move the `NuGetConfigFileReader` to its own file for more overview
